requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
numpy
scikit-learn
xgboost
ta
ccxt
psycopg2-binary
sqlalchemy
